# web-contract/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Trading terminal API",
    "version": "0.1.0"
  },
...
  "paths": {
...
    "/api/v1/signals/history": {
      "get": {
        "summary": "State transitions (live engine only; drills are never recorded)",
        "tags": [
          "signals"
        ],
        "parameters": [
          {
            "schema": {
              "default": 50,
              "type": "integer",
              "minimum": 1,
              "maximum": 200
            },
            "in": "query",
            "name": "limit",
            "required": false
          },
          {
            "schema": {
              "type": "integer",
              "minimum": -9007199254740991,
              "maximum": 9007199254740991
            },
            "in": "query",
            "name": "cursor",
            "required": false
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "asOf": {
                      "type": "string"
                    },
                    "mode": {
                      "type": "string",
                      "enum": [
                        "live",
                        "replay",
                        "drill",
                        "offline"
                      ]
                    },
                    "source": {
                      "type": "string"
                    },
                    "items": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "number"
                          },
                          "ts": {
                            "type": "string"
                          },
                          "state": {
                            "type": "string"
                          },
                          "mode": {
                            "type": "string"
                          },
                          "payload": {}
                        },
                        "required": [
                          "id",
                          "ts",
                          "state",
                          "mode",
                          "payload"
                        ],
                        "additionalProperties": false
                      }
                    },
                    "nextCursor": {
                      "nullable": true,
                      "type": "number"
                    }
                  },
                  "required": [
                    "asOf",
                    "mode",
                    "source",
                    "items",
                    "nextCursor"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/demo/drill": {
      "post": {
        "summary": "Inject a synthetic 6.5% crash + ~$40M liquidation burst + rebound into an isolated copy of state",
        "tags": [
          "demo"
        ],
        "description": "Every message is tagged `synthetic: true` and `mode: \"drill\"`. Real data, accounts and ledgers are never touched. Returns immediately; poll GET /demo/drill.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "default": {
                  "speed": 1
                },
                "type": "object",
                "properties": {
                  "speed": {
                    "default": 1,
                    "type": "number",
                    "minimum": 0.5,
                    "maximum": 1000
                  }
                }
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "enum": [
                        "started"
                      ]
                    },
                    "speed": {
                      "type": "number"
                    }
                  },
                  "required": [
                    "status",
                    "speed"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      },
      "get": {
        "summary": "Status and result of the latest drill",
        "tags": [
          "demo"
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {}
                }
              }
            }
          }
        }
      }
    }
  },
  "servers": [
    {
      "url": "/"
    }
  ]
}